COPY communicates with helpers through non-blocking pipe
This is an alternative solution for #507 with non-blocking call to open pipe.

Gprestore part is not changed. If error occures and we have
--on-error-continue, gprestore will create skip_oid file on every
segment to inform gpbackup-helpers to also skip the table.

The difference is this solution doesn't have additional concurrency
on the writer (gpbackup-helper) to check for skip file and try
to open the pipe in parallel.

Instead, the writer will try to open the pipe and if gets back
the specific syscall.ENXIO error (meaning that the reader
(COPY command) is not ready to consume from it yet).

Because this could also mean that gprestore decides to skip
the table (error on restore + --on-error-continue scenario),
when getting this error writer will also check for skips file
and skip the table if discovers it.

Authored-by: Kate Dontsova kdontsova@vmware.com