Queue parallelized COPY commands for gpbackup_helper to consume
To reduce the amount of time gpbackup_helper is spent blocked waiting
for data to copy, gpbackup and gprestore can now queue COPY commands to
speed up --single-data-file backups and restores if the database has cpu
cycles to spare.

Added flag --single-data-file-copy-prefetch (Default 1). Can only be
used with --single-data-file

This can improve backup and restore times significantly for databases
that contain a lot of tables with very little data where the majority of
time is spent waiting for gpdb to initiate COPY.

Created scale tests for --single-data-file-copy-prefetch flag

Resolve issue where COPY commands may not be terminated properly when a
SIGINT signal is received.

Co-authored-by: Kevin Yeap <kyeap@vmware.com>
Co-authored-by: Brent Doil <bdoil@vmware.com>