Improve name handling for table filters
Our previous processing could not handle dots in table names, and was
not escaping double quotes if a user included them. This commit handling
should achieve parity with what pg_dump is able to process.

In addition, the input filter lists were re-processed in multiple places
throughout the codebase, causing some room for quote_ident to possibly
be run on the same name more than once.  To clean things up, we now use
a struct representation that is populated one time, and then referenced
for the remainder of the run.  We now process based on OID wherever the
filter lists are needed.  This required fairly wide-ranging changes to
the integration test suite.

This commit leaves one TODO that should be addressed.  Our options
module needs access to this struct, but cannot import the backups module
due to circular import errors.  This presents a challenge, and the
current solution is to have identical options.Relation and
backup.Relation structs, and a routine to convert from one to the other.
This is obviously inefficient, but the alternative would be to pull
large chunks of logic "upstream" into options so that options can be the
official home of this struct, and that seems very invasive to the
codebase.  A better solution would be good, but none seems clear at the
moment.