Integration: fix example plugin integration test to use relative paths
When gpbackup repo is not in $GOPATH and integration tests are run,
tests that use the example plugin will hang because the test cannot
proceed without the example plugin reading data from test pipes. This
happens because the example plugin and it's config file locations were
hardcoded to be in $GOPATH. This is a legacy issue from when go projects
were required to be in $GOPATH. This requirement was changed to be
optional in go1.11. This is an update to look for example plugin using a
relative path. The test plugin config file is also now setup before
tests run because the executablepath is also updated to be a relative
path.