Add --no-privileges flag to gpbackup
This flag causes gpbackup to omit statements related to object ownership
and privileges from the metadata file, in order to (among other things)
allow a predata restore to be done without needing all of the original
roles to exist in the target database or needing to perform a restore of
globals beforehand.

This commit does not add the same flag to gprestore to allow ignoring
ownership and privilege statements for objects during a restore that
does include them, but it does refactor the ToC logic to assign those
statements their own types (as they are currently treated as tables) to
enable that functionality to be more easily added to gprestore in a
later commit.

Original commit is written by Jamie.

Co-authored-by: Jamie McAtamney <jmcatamney@vmware.com>
Co-authored-by: Kevin Yeap <kyeap@vmware.com>