Remove redundant GetBackupDataSet unit tests
The only caller of GetBackupDataSet is DoBackup, which now tests
whether --section=data is set before calling the function. This
change is to avoid a global variable check within the function.
Its job is to remove external/foreign tables, and now it does
just that. Remove the now unnecessary unit tests.